[FLINK-5795][TableAPI&SQL] add functionIdentifier method.
