+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/plan/logical/operators.scala
@@ -694,7 +694,7 @@ case class LogicalTableFunctionCall(
     val function = new FlinkTableFunctionImpl(resultType, fieldIndexes, fieldNames, evalMethod)
     val typeFactory = relBuilder.getTypeFactory.asInstanceOf[FlinkTypeFactory]
     val sqlFunction = TableSqlFunction(
-      tableFunction.toString,
+      tableFunction.functionIdentifier,
       tableFunction,
       resultType,
       typeFactory,
